Create (C) Monsters (easy version).cpp
contest: Codeforces Round #850 (Div. 2, based on VK Cup 2022 - Final Round)
problem: (C) Monsters (easy version)
note to myself: use long long instead of int if you are calculating the sum.
note: 
the goal is to make the array increasing by 1 or 0, 
the objective is to find the min no. of moves needed for it.